Move types into index.ts (BETA) (#178)
