+++ lib/index.ts
@@ -1,6 +1,24 @@
 import base64_url_decode from "./base64_url_decode";
-import { JwtDecodeOptions, JwtHeader, JwtPayload } from "./global";
-export * from './global';
+
+export interface JwtDecodeOptions {
+  header?: boolean;
+}
+
+export interface JwtHeader {
+  typ?: string;
+  alg?: string;
+  kid?: string;
+}
+
+export interface JwtPayload {
+  iss?: string;
+  sub?: string;
+  aud?: string[] | string;
+  exp?: number;
+  nbf?: number;
+  iat?: number;
+  jti?: string;
+}
 
 export class InvalidTokenError extends Error {
   constructor(message: string) {
